Add Ts support to Load test
