+++ .env
@@ -1,0 +1,10 @@
+APPINSIGHTS_INSTRUMENTATIONKEY=foo
+rate=20
+duration=1m
+preAllocatedVUs=10
+maxVUs=100
+PRODUCER_BASE_URL=https://io-d-reminder-producer.azurewebsites.net
+PU_BASE_URL=https://api-app.internal.io.pagopa.it
+API_SUBSCRIPTION_KEY=a6c4354aa2a04192b7ce4f0f63caf833
+API_ENVIRONMENT=weu-prod01
+SCRIPT=message_payments
