+++ package.json
@@ -1,0 +1,63 @@
+{
+    "name": "@pagopa/io-payment-updater-load-test",
+    "description": "",
+    "author": "IO team",
+    "repository": "https://github.com/pagopa/io-payment-updater-load-test",
+    "version": "1.0.0",
+    "license": "MIT",
+    "scripts": {
+      "prebuild": "shx rm -rf dist",
+      "postbuild": "dependency-check package.json --no-dev --missing ./dist/**/*.js",
+      "build": "tsc",
+      "watch": "tsc -w",
+      "dev:message-payments": "shx rm -rf dist && webpack && docker run --env-file .env -i --rm -v $(pwd)/dist:/dist loadimpact/k6 run /dist/message_payments.js",
+      "start": "shx rm -rf dist && webpack && docker run -e PRODUCER_BASE_URL=${PRODUCER_BASE_URL} -e PU_BASE_URL=${PU_BASE_URL} -e API_ENVIRONMENT=${API_ENVIRONMENT} -e -e API_SUBSCRIPTION_KEY=${API_SUBSCRIPTION_KEY} -e rate=${rate} -e duration=${duration} -e preAllocatedVUs=${preAllocatedVUs} -e maxVUs=${maxVUs} loadimpact/k6 run /dist/message_payments.js",
+      "predev": "npm-run-all build",
+      "dev": "npm-run-all --parallel start watch",
+      "preversion": "auto-changelog  --config .auto-changelog.json --unreleased --commit-limit false --stdout --template preview.hbs",
+      "version": "auto-changelog -p --config .auto-changelog.json --unreleased && git add CHANGELOG.md",
+      "webpack": "shx rm -rf dist && webpack"
+    },
+    "devDependencies": {
+        "@babel/core": "7.13.16",
+        "@babel/plugin-proposal-class-properties": "7.13.0",
+        "@babel/plugin-proposal-object-rest-spread": "7.13.8",
+        "@babel/preset-env": "7.13.15",
+        "@babel/preset-typescript": "7.13.0",
+        "@types/webpack": "5.28.0",
+        "@pagopa/danger-custom-rules": "^2.0.3",
+        "@pagopa/eslint-config": "^1.3.1",
+        "@pagopa/openapi-codegen-ts": "^11.1.0",
+        "@types/k6": "^0.37.1",
+        "babel-loader": "8.2.2",
+        "clean-webpack-plugin": "4.0.0-alpha.0",
+        "copy-webpack-plugin": "^9.0.1",
+        "webpack": "5.35.1",
+        "webpack-cli": "4.6.0",
+        "webpack-glob-entries": "^1.0.1",
+        "danger": "^10.3.0",
+        "dependency-check": "^4.1.0",
+        "dotenv-cli": "^3.1.0",
+        "eslint-plugin-prettier": "^3.3.1",
+        "modclean": "^3.0.0-beta.1",
+        "npm-run-all": "^4.1.5",
+        "oval": "^1.0.0",
+        "prettier": "^1.18.2",
+        "shx": "^0.3.2",
+        "typescript": "^4.3.5"
+    },
+    "dependencies": {
+        "@pagopa/ts-commons": "^10.7.1",
+        "k6": "^0.0.0",
+        "k6-stat": "1.0.6",
+        "applicationinsights": "^1.7.4",
+        "fp-ts": "^2.10.5",
+        "io-ts": "^2.2.16",
+        "lodash": "^4.17.21",
+        "node-fetch": "^2.6.7"
+    },
+    "resolutions": {
+      "handlebars": "~4.5.3"
+    }
+  }
+
