+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+      "target": "ES2015",
+      "rootDir": "src",
+      "outDir": "dist",
+      "module": "commonjs",
+      "moduleResolution": "node",
+      "strict": true,
+      "sourceMap": false,
+      "resolveJsonModule": true,
+      "skipLibCheck": true
+    },
+    "exclude": [
+      "__mocks__",
+      "dist",
+      "node_modules",
+      "**/__tests__/*",
+      "Dangerfile.ts"
+    ]
+  }
